Use int instead of ssize_t parameter for shift_lines() function
Strictly speaking, ssize_t is for "either a count of bytes or an error
indicator (-1)", not for a count of bytes with arbitrary use of the
sign bit.

+++ src/edit.h
@@ -3,7 +3,6 @@
 
 #include <stdbool.h>
 #include <stddef.h>
-#include <sys/types.h>
 #include "util/unicode.h"
 
 void select_block(void);
@@ -16,7 +15,7 @@ void delete_ch(void);
 void erase(void);
 void insert_ch(CodePoint ch);
 void join_lines(void);
-void shift_lines(ssize_t count);
+void shift_lines(int count);
 void clear_lines(void);
 void new_line(void);
 void format_paragraph(int pw);
